tests/requirements.txt
# tests/requirements.txt — BMT AI OS test dependencies
# Install with: pip install -r tests/requirements.txt

# Test framework
pytest>=8.0,<9.0
pytest-timeout>=2.2,<3.0
pytest-html>=4.0,<5.0

# HTTP assertions for service health checks
requests>=2.31,<3.0

# SSH into QEMU guest for integration tests
paramiko>=3.4,<4.0

# YAML parsing for docker-compose validation
pyyaml>=6.0,<7.0

# Docker SDK for compose smoke tests
docker>=7.0,<8.0

# Main package (needed by unit tests that import bmt_ai_os)
-e .
